# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API EBAC",
        "description": "Documentation OpenAPI des API EBAC : églises, paramètres académiques et pré-inscription.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Serveur courant"
        },
        {
            "url": "https://api-ebac.severinzran.ci/api/v1",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/parametres/modules": {
            "get": {
                "tags": [
                    "Modules"
                ],
                "operationId": "listerModules",
                "parameters": [
                    {
                        "name": "matiere",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "niveau",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "ordre",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste filtrée des modules"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Modules"
                ],
                "operationId": "creerModule",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ModulePayload"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Module créé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/parametres/modules/{id}": {
            "get": {
                "tags": [
                    "Modules"
                ],
                "operationId": "afficherModule",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Module trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Modules"
                ],
                "operationId": "remplacerModule",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ModulePayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Module modifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Modules"
                ],
                "operationId": "supprimerModule",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Module supprimé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Modules"
                ],
                "operationId": "modifierModule",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ModulePayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Module modifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ModulePayload": {
                "required": [
                    "id_matiere",
                    "libelle"
                ],
                "properties": {
                    "id_matiere": {
                        "type": "integer"
                    },
                    "code": {
                        "type": "string",
                        "nullable": true
                    },
                    "libelle": {
                        "type": "string"
                    },
                    "ordre": {
                        "type": "integer",
                        "default": 1
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Jeton retourné après la connexion et la confirmation OTP.",
                "bearerFormat": "Token Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}